u11/cp: Modify the cellinfo data type，because cp cellinfo data type is int
U11-229

+++ at_client/at_api.h
@@ -145,7 +145,7 @@ typedef struct
   CURR_MODE curr_mod;
   DUPLEX_MODE duplex_mode;
   int ue_category;
-  unsigned short cellId;
+  unsigned int cellId;
   unsigned int lacId;
   int rsrp;
   short rsrq;
